backend/requirements.txt
# FastAPI
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1

# RAG / AI
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.19
openai==1.12.0
tiktoken==0.5.2

# Vector DB
qdrant-client==1.7.1

# Document processing
pypdf2==3.0.1
beautifulsoup4==4.12.3
python-docx==1.1.0
pandas==2.2.0

# Utilities
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.26.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Dev
pytest==8.0.0
pytest-asyncio==0.23.4
